+++ examples/all_of/column_models.py
@@ -1,0 +1,3 @@
+from open_alchemy import init_yaml
+
+init_yaml("column-example-spec.yml", models_filename="column_models_auto.py")