+++ examples/composite_index/models.py
@@ -1,0 +1,3 @@
+from open_alchemy import init_yaml
+
+init_yaml("example-spec.yml", models_filename="models_auto.py")