+++ examples/composite_index_models.py
@@ -1,5 +1,0 @@
-from open_alchemy import init_yaml
-
-init_yaml(
-    "composite-index-example-spec.yml", models_filename="composite_index_models_auto.py"
-)