+++ examples/relationship/many_to_one/custom_foreign_key_models.py
@@ -1,0 +1,6 @@
+from open_alchemy import init_yaml
+
+init_yaml(
+    "custom-foreign-key-example-spec.yml",
+    models_filename="custom_foreign_key_models_auto.py",
+)