+++ examples/relationship_one_to_one_models.py
@@ -1,6 +1,0 @@
-from open_alchemy import init_yaml
-
-init_yaml(
-    "relationship-one-to-one-example-spec.yml",
-    models_filename="relationship_one_to_one_models_auto.py",
-)