+++ examples/simple_models.py
@@ -1,3 +1,0 @@
-from open_alchemy import init_yaml
-
-init_yaml("simple-example-spec.yml", models_filename="simple_models_auto.py")