+++ examples/unique_constraint_models.py
@@ -1,6 +1,0 @@
-from open_alchemy import init_yaml
-
-init_yaml(
-    "unique-constraint-example-spec.yml",
-    models_filename="unique_constraint_models_auto.py",
-)